Fixed copr API URL
The old url was not the correct one, if one would try to login to get his API key, he would be redirected to the homepage of the new site. In fact I have the impression that the entire URL should be using a HTTP redirect to the new domain.