Include page deprecation notice for MongoDB.
While the page was updated to contain some working approach,
installing newer versions on current Fedora Linuxes with simple DNF
install seems not possible. Therefore the page is being deprecated
and will be remove in the future. This deprecation will be accompanied
by an email to fedora-devel calling for contributors that would
update the content to a better state.