[Rust] Expand section about cargo + crate dependencies
Rust crates as packaged by Fedora are not really a suitable development target - this is one of the reasons why using them as a replacement for crates.io is not a supported use case.

While it was intentional that there was no documentation for how to use the Fedora packages for Rust crates for local development (it is not a supported use case, after all), this has caused confusion on multiple occasions. To remove this ambiguity, this commit explicitly documents this use case as unsupported in the developer documentation.